# api/swagger/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для маркетплейса на Go с JWT аутентификацией, управлением магазинами, продуктами и категориями",
        "title": "Go Shop API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Go Shop API Support",
            "url": "https://github.com/ReversIncorp/go-shop",
            "email": "a.savko.developer@goshop.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/users/login": {
            "post": {
                "description": "Аутентифицирует пользователя и возвращает токены доступа",
                "tags": [
                    "users"
                ],
                "summary": "Вход пользователя",
                "parameters": [
                    {
                        "description": "Данные для входа",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.LoginCredentials"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный вход",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Неверные учетные данные",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    }
                }
            }
        },
        "/users/logout": {
            "post": {
                "description": "Завершает сессию пользователя",
                "tags": [
                    "users"
                ],
                "summary": "Выход из системы",
                "parameters": [
                    {
                        "description": "Access token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный выход",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Ошибка при выходе",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    }
                }
            }
        },
        "/users/refresh": {
            "post": {
                "description": "Обновляет токены доступа используя refresh token",
                "tags": [
                    "users"
                ],
                "summary": "Обновление сессии",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Токены обновлены",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Неверный refresh token",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    }
                }
            }
        },
        "/users/register": {
            "post": {
                "description": "Создает нового пользователя в системе",
                "tags": [
                    "users"
                ],
                "summary": "Регистрация пользователя",
                "parameters": [
                    {
                        "description": "Данные пользователя",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Пользователь успешно зарегистрирован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "Возвращает информацию о пользователе по его ID",
                "tags": [
                    "users"
                ],
                "summary": "Получение пользователя по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о пользователе",
                        "schema": {
                            "$ref": "#/definitions/entities.User"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "entities.LoginCredentials": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
...
        "entities.User": {
            "type": "object",
            "required": [
                "email",
                "is_seller",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_seller": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "errorhandling.ResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP статус-код",
                    "type": "integer"
                },
                "details": {
                    "description": "Сообщение для пользователя",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите \"Bearer\" за которым следует пробел и JWT токен.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}